backend/requirements/base.txt
Django==4.2.11
djangorestframework==3.14.0
django-filter==24.1
django-cors-headers==4.3.1

psycopg2-binary==2.9.9

celery==5.3.4
redis==5.0.1
kombu==5.3.5

requests==2.31.0
beautifulsoup4==4.12.2
pdfplumber==0.10.3

scrapy==2.11.0
playwright==1.42.0
scrapy-playwright==0.0.21

python-dateutil==2.8.2
pytz==2024.1
